+++ www/talsec.ts
@@ -9,6 +9,8 @@ export interface Talsec {
   ) => Promise<void>;
   addToWhitelist: (packageName: string) => Promise<string>;
   getAppIcon: (packageName: string) => Promise<string>;
+  blockScreenCapture: (enable: boolean) => Promise<string>;
+  isScreenCaptureBlocked: () => Promise<boolean>;
 }
 
 export type SuspiciousAppInfo = {
@@ -40,6 +42,8 @@ export type NativeEventEmitterActions = {
   systemVPN?: () => any;
   malware?: (suspiciousApps: SuspiciousAppInfo[]) => any;
   adbEnabled?: () => any;
+  screenshot?: () => any;
+  screenRecording?: () => any;
 };
 
 export type TalsecConfig = {
@@ -85,6 +89,8 @@ class Threat {
   static DevMode = new Threat(0);
   static Malware = new Threat(0);
   static ADBEnabled = new Threat(0);
+  static Screenshot = new Threat(0);
+  static ScreenRecording = new Threat(0);
 
   constructor(value: number) {
     this.value = value;
@@ -107,6 +113,8 @@ class Threat {
           this.DevMode,
           this.Malware,
           this.ADBEnabled,
+          this.Screenshot,
+          this.ScreenRecording,
         ]
       : [
           this.AppIntegrity,
@@ -120,9 +128,16 @@ class Threat {
           this.DeviceBinding,
           this.DeviceID,
           this.UnofficialStore,
+          this.Screenshot,
+          this.ScreenRecording,
         ];
   };
 }
+
+const ScreenCaptureStatus = {
+  ALLOWED: 0,
+  BLOCKED: 1,
+} as const;
 
 const getThreatCount = (): number => {
   return Threat.getValues().length;
@@ -240,6 +255,12 @@ const start = async (
       case Threat.ADBEnabled.value:
         eventListenerConfig.adbEnabled?.();
         break;
+      case Threat.Screenshot.value:
+        eventListenerConfig.screenshot?.();
+        break;
+      case Threat.ScreenRecording.value:
+        eventListenerConfig.screenRecording?.();
+        break;
       default:
         onInvalidCallback();
         break;
@@ -286,9 +307,43 @@ const getAppIcon = (packageName: string): Promise<string> => {
   });
 };
 
+const blockScreenCapture = (enable: boolean): Promise<string> => {
+  if (cordova.platformId === 'ios') {
+    return Promise.reject(
+      'Blocking/Unblocking Screen Capture is not available on iOS',
+    );
+  }
+
+  return new Promise((resolve, reject) => {
+    cordova.exec(resolve, reject, 'TalsecPlugin', 'blockScreenCapture', [
+      enable,
+    ]);
+  });
+};
+
+const isScreenCaptureBlocked = (): Promise<boolean> => {
+  if (cordova.platformId === 'ios') {
+    return Promise.reject(
+      'Checking Screen Capture Status is not available on iOS',
+    );
+  }
+
+  return new Promise((resolve, reject) => {
+    cordova.exec(
+      (result: number) => resolve(result === ScreenCaptureStatus.BLOCKED),
+      reject,
+      'TalsecPlugin',
+      'isScreenCaptureBlocked',
+      [],
+    );
+  });
+};
+
 // @ts-ignore
 module.exports = {
   start,
   addToWhitelist,
   getAppIcon,
+  blockScreenCapture,
+  isScreenCaptureBlocked,
 };